% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICS2014Ch.R
\docType{data}
\name{MICS2014Ch}
\alias{MICS2014Ch}
\title{Multiple Indicator Cluster Survey (MICS) 2014 Child Questionnaire Data for Punjab, Pakistan}
\format{A data frame with 31083 rows and 270 variables:
\describe{
  \item{\code{HH1}}{Cluster number}
  \item{\code{HH2}}{Household number}
  \item{\code{LN}}{Line number}
  \item{\code{UF1}}{Cluster number}
  \item{\code{UF2}}{Household number}
  \item{\code{UF4}}{Child's line number}
  \item{\code{UF6}}{Mother / Caretaker's line number}
  \item{\code{UF7}}{Interviewer number}
  \item{\code{UF8D}}{Day of interview}
  \item{\code{UF8M}}{Month of interview}
  \item{\code{UF8Y}}{Year of interview}
  \item{\code{UF9}}{Result of interview for children under 5}
  \item{\code{UF10}}{Field editor}
  \item{\code{UF11}}{Main data entry clerk}
  \item{\code{UF12H}}{Start of interview - Hour}
  \item{\code{UF12M}}{Start of interview - Minutes}
  \item{\code{UF13H}}{End of interview - Hour}
  \item{\code{UF13M}}{End of interview - Minutes}
  \item{\code{AG1D}}{Day of birth of child}
  \item{\code{AG1M}}{Month of birth of child}
  \item{\code{AG1Y}}{Year of birth of child}
  \item{\code{AG2}}{Age of child}
  \item{\code{BR1}}{Birth certificate}
  \item{\code{BR2}}{Birth registered}
  \item{\code{BR3}}{Know how to register birth}
  \item{\code{EC1}}{Number of children's books or picture books for child}
  \item{\code{EC2A}}{Homemade toys}
  \item{\code{EC2B}}{Toys from shops}
  \item{\code{EC2C}}{Household objects or outside objects}
  \item{\code{EC3A}}{In past week, days left alone for more than 1 hour}
  \item{\code{EC3B}}{In past week, days left with other child for more than 1 hour}
  \item{\code{EC5}}{Attends early childhood education programme}
  \item{\code{EC7AA}}{Read books - Mother}
  \item{\code{EC7AB}}{Read books - Father}
  \item{\code{EC7AX}}{Read books - Other}
  \item{\code{EC7AY}}{Read books - No one}
  \item{\code{EC7BA}}{Tell stories - Mother}
  \item{\code{EC7BB}}{Tell stories - Father}
  \item{\code{EC7BX}}{Tell stories - Other}
  \item{\code{EC7BY}}{Tell stories - No one}
  \item{\code{EC7CA}}{Sang songs - Mother}
  \item{\code{EC7CB}}{Sang songs - Father}
  \item{\code{EC7CX}}{Sang songs - Other}
  \item{\code{EC7CY}}{Sang songs - No one}
  \item{\code{EC7DA}}{Took outside - Mother}
  \item{\code{EC7DB}}{Took outside - Father}
  \item{\code{EC7DX}}{Took outside - Other}
  \item{\code{EC7DY}}{Took outside - No one}
  \item{\code{EC7EA}}{Played with - Mother}
  \item{\code{EC7EB}}{Played with - Father}
  \item{\code{EC7EX}}{Played with - Other}
  \item{\code{EC7EY}}{Played with - No one}
  \item{\code{EC7FA}}{Named/counted - Mother}
  \item{\code{EC7FB}}{Named/counted - Father}
  \item{\code{EC7FX}}{Named/counted - Other}
  \item{\code{EC7FY}}{Named/counted - No one}
  \item{\code{EC8}}{Child identifies at least ten letters of the alphabet}
  \item{\code{EC9}}{Child reads at least four simple, popular words}
  \item{\code{EC10}}{Child knows name and recognizes symbol of all numbers from 1-10}
  \item{\code{EC11}}{Child able to pick up small object with 2 fingers}
  \item{\code{EC12}}{Child sometimes too sick to play}
  \item{\code{EC13}}{Child follows simple directions}
  \item{\code{EC14}}{Child able to do something independently}
  \item{\code{EC15}}{Child gets along well with other children}
  \item{\code{EC16}}{Child kicks, bites or hits other children or adults}
  \item{\code{EC17}}{Child gets distracted easily}
  \item{\code{BD2}}{Child ever been breastfed}
  \item{\code{BD3}}{Child still being breastfed}
  \item{\code{BD4}}{Child drank anything else from the bottle with a nipple yesterday}
  \item{\code{BD5}}{Child drank ORS yesterday}
  \item{\code{BD6}}{Child drank or ate vitamin or mineral supplements yesterday}
  \item{\code{BD7A}}{Child drank plain water yesterday}
  \item{\code{BD7B}}{Child drank juice or juice drinks yesterday}
  \item{\code{BD7C}}{Child drank clear soup (any type) yesterday}
  \item{\code{BD7D}}{Child drank milk such as tinned, powdered, fresh animal milk or Cured shake (Lasi) yesterday}
  \item{\code{BD7DN}}{Times child drank milk}
  \item{\code{BD7E}}{Child drank infant formula yesterday}
  \item{\code{BD7EN}}{Times child drank infant formula}
  \item{\code{BD7F}}{Child drank any other liquid yesterday}
  \item{\code{BD8A}}{Child drank or ate yogurt yesterday}
  \item{\code{BD8AN}}{Times drank or ate yogurt}
  \item{\code{BD8B}}{Child ate cerelac or any fortified baby food}
  \item{\code{BD8C}}{Child ate foods made from grains (Dalia etc)}
  \item{\code{BD8D}}{Child ate pumpkin, carrots, squash etc.}
  \item{\code{BD8E}}{Child ate white potatoes, white yams, manioc, turnip, spinach, green beans  etc.}
  \item{\code{BD8F}}{Child ate green leafy vegetables}
  \item{\code{BD8G}}{Child ate ripe mangoes, papayas}
  \item{\code{BD8H}}{Child ate other fruits or vegetables}
  \item{\code{BD8I}}{Child ate liver, kidney, heart or other organ meat}
  \item{\code{BD8J}}{Child ate meat, such as beef, pork, lamb, goat, chicken, duck}
  \item{\code{BD8K}}{Child ate eggs}
  \item{\code{BD8L}}{Child ate fresh or dried fish or shellfish}
  \item{\code{BD8M}}{Child ate beans, lentils or nuts}
  \item{\code{BD8N}}{Child ate cheese or other food made from milk}
  \item{\code{BD8O}}{Child ate other solid, semi-solid or soft food yesterday}
  \item{\code{BD11}}{Times child ate solid or semi-solid food}
  \item{\code{IM1}}{Vaccination card for child}
  \item{\code{IM2}}{Ever had vaccination card}
  \item{\code{IM3BD}}{Day of BCG immunization}
  \item{\code{IM3BM}}{Month of BCG immunization}
  \item{\code{IM3BY}}{Year of BCG immunization}
  \item{\code{IM3P0D}}{Day of Polio at birth immunization}
  \item{\code{IM3P0M}}{Month of Polio at birth immunization}
  \item{\code{IM3P0Y}}{Year of Polio at birth immunization}
  \item{\code{IM3P1D}}{Day of Polio1 immunization}
  \item{\code{IM3P1M}}{Month of Polio1 immunization}
  \item{\code{IM3P1Y}}{Year of Polio1 immunization}
  \item{\code{IM3P2D}}{Day of Polio2 immunization}
  \item{\code{IM3P2M}}{Month of Polio2 immunization}
  \item{\code{IM3P2Y}}{Year of Polio2 immunization}
  \item{\code{IM3P3D}}{Day of Polio3 immunization}
  \item{\code{IM3P3M}}{Month of Polio3 immunization}
  \item{\code{IM3P3Y}}{Year of Polio3 immunization}
  \item{\code{IM3T1D}}{Day of PENTA1 immunization}
  \item{\code{IM3T1M}}{Month of PENTA1 immunization}
  \item{\code{IM3T1Y}}{Year of PENTA1 immunization}
  \item{\code{IM3T2D}}{Day of PENTA2 immunization}
  \item{\code{IM3T2M}}{Month of PENTA2 immunization}
  \item{\code{IM3T2Y}}{Year of PENTA2 immunization}
  \item{\code{IM3T3D}}{Day of PENTA3 immunization}
  \item{\code{IM3T3M}}{Month of PENTA3 immunization}
  \item{\code{IM3T3Y}}{Year of PENTA3 immunization}
  \item{\code{IM3M1D}}{Day measles1 or MMR immunization}
  \item{\code{IM3M1M}}{Month Measles1 or MMR immunization}
  \item{\code{IM3M1Y}}{Year of Measles1 or MMR immunization}
  \item{\code{IM3M2D}}{Day Measles2 or MMR immunization}
  \item{\code{IM3M2M}}{Month Measles2 or MMR immunization}
  \item{\code{IM3M2Y}}{Year Measles2 or MMR immunization}
  \item{\code{IM5}}{Child received any other vaccinations}
  \item{\code{IM6}}{Child ever received any vaccinations}
  \item{\code{IM7}}{Child ever given BCG vaccination}
  \item{\code{IM8}}{Child ever given Polio vaccination}
  \item{\code{IM9}}{Polio first given just after birth or later}
  \item{\code{IM10}}{Times child given Polio vaccination}
  \item{\code{IM11}}{Child ever given DPT vaccination}
  \item{\code{IM12}}{Times child given DPT vaccination}
  \item{\code{IM15A}}{Child ever given Pentavalent vaccination}
  \item{\code{IM15B}}{Times child given Pentavalent vaccination}
  \item{\code{IM16}}{Child ever given Measles or MMR vaccination}
  \item{\code{IM19A}}{Child participated in: Anti-Polio campaign - NID (-------------}
  \item{\code{IM19B}}{Child participated in: Mother and Child week (---------------)}
  \item{\code{IM19C}}{Child participated in: Child Health Day (---------------------)}
  \item{\code{VS1}}{Child received a Vitamin A dose within the last 6 months}
  \item{\code{CA1}}{Child had diarrhoea in last 2 weeks}
  \item{\code{CA2}}{Child drank less or more during illness}
  \item{\code{CA3}}{Child ate less or more during illness}
  \item{\code{CA3A}}{Sought advice or treatment for the illness from any source}
  \item{\code{CA3BA}}{Place sought care: (public sector) Government hospital}
  \item{\code{CA3BB}}{Place sought care: (public sector) Government health center}
  \item{\code{CA3BC}}{Place sought care: (public sector) Government health post/dispensary}
  \item{\code{CA3BD}}{Place sought care: (public sector) Lady health worker (LHW)}
  \item{\code{CA3BE}}{Place sought care: (public sector) Mobile / Outreach clinic}
  \item{\code{CA3BH}}{Place sought care: Other public}
  \item{\code{CA3BI}}{Place sought care: Private hospital / clinic}
  \item{\code{CA3BJ}}{Place sought care: Private physician}
  \item{\code{CA3BK}}{Place sought care: Private pharmacy}
  \item{\code{CA3BL}}{Place sought care: Mobile clinic}
  \item{\code{CA3BO}}{Place sought care: Other private medical}
  \item{\code{CA3BP}}{Place sought care: Relative / Friend}
  \item{\code{CA3BQ}}{Place sought care: Shop}
  \item{\code{CA3BR}}{Place sought care: Traditional practitioner}
  \item{\code{CA3BX}}{Place sought care: Other}
  \item{\code{CA4A}}{Drank fluid made from special packet (ORS)}
  \item{\code{CA4B}}{Pre-packaged ORS fluid for diarrhoea}
  \item{\code{CA4BB}}{Place where ORS was obtained}
  \item{\code{CA4CA}}{Child given zinc tablets}
  \item{\code{CA4CB}}{Child given zinc syrup}
  \item{\code{CA4E}}{Place where zinc was obtained}
  \item{\code{CA4FA}}{Child given home made fluid (boiled water, sugar and salt)}
  \item{\code{CA4FB}}{Child given home made fluid (Others - specify)}
  \item{\code{CA5}}{Anything else given to treat the diarrhoea}
  \item{\code{CA6A}}{Other treatment (pill or syrup): Antibiotic}
  \item{\code{CA6B}}{Other treatment (pill or syrup): Antimotility}
  \item{\code{CA6G}}{Other treatment (pill or syrup): Other (not antibiotic, antimotility or zinc)}
  \item{\code{CA6H}}{Other treatment (pill or syrup): Unknown}
  \item{\code{CA6L}}{Other treatment (injection): Antibiotic}
  \item{\code{CA6M}}{Other treatment (injection): Non-antibiotic}
  \item{\code{CA6N}}{Other treatment (injection): Unknown}
  \item{\code{CA6O}}{Other treatment: Intravenous}
  \item{\code{CA6Q}}{Other treatment: Home remedy/Herbal medicine}
  \item{\code{CA6X}}{Other treatment: Other}
  \item{\code{CA6AA}}{Child ill with fever in last 2 weeks}
  \item{\code{CA6BB}}{During fever blood was taken from his/her finger of heel}
  \item{\code{CA7}}{Child ill with cough in last 2 weeks}
  \item{\code{CA8}}{Difficulty breathing during illness with cough}
  \item{\code{CA9}}{Symptoms due to problem in chest or blocked nose}
  \item{\code{CA10}}{Sought advice or treatment for illness}
  \item{\code{CA11A}}{Place sought care: (public sector) Government hospital}
  \item{\code{CA11B}}{Place sought care: (public sector) Government health center}
  \item{\code{CA11C}}{Place sought care: (public sector) Government health post/dispensary}
  \item{\code{CA11D}}{Place sought care: (public sector) Lady health worker (LHW)}
  \item{\code{CA11E}}{Place sought care: (public sector) Mobile / Outreach clinic}
  \item{\code{CA11H}}{Place sought care: Other public}
  \item{\code{CA11I}}{Place sought care: Private hospital / clinic}
  \item{\code{CA11J}}{Place sought care: Private physician}
  \item{\code{CA11K}}{Place sought care: Private pharmacy}
  \item{\code{CA11L}}{Place sought care: Mobile clinic}
  \item{\code{CA11O}}{Place sought care: Other private medical}
  \item{\code{CA11P}}{Place sought care: Relative / Friend}
  \item{\code{CA11Q}}{Place sought care: Shop}
  \item{\code{CA11R}}{Place sought care: Traditional practitioner}
  \item{\code{CA11X}}{Place sought care: Other}
  \item{\code{CA12}}{Given medicine to treat this illness}
  \item{\code{CA13A}}{Medicine: SP / Fansidar}
  \item{\code{CA13B}}{Medicine: Chloroquine}
  \item{\code{CA13C}}{Medicine: Amodiaquine}
  \item{\code{CA13D}}{Medicine: Quinine}
  \item{\code{CA13E}}{Medicine: Combiniation with Artemisinin}
  \item{\code{CA13H}}{Medicine: Other anti-malarial}
  \item{\code{CA13I}}{Medicine: Antibiotic pill / syrup}
  \item{\code{CA13J}}{Medicine: Antibiotic injection}
  \item{\code{CA13P}}{Medicine: Paracetamol / Panadol / Acetaminophen}
  \item{\code{CA13Q}}{Medicine: Aspirin}
  \item{\code{CA13R}}{Medicine: Ibupropfen}
  \item{\code{CA13X}}{Medicine: Other}
  \item{\code{CA13Z}}{Medicine: DK}
  \item{\code{CA13BB}}{Place where antibiotics were obtained}
  \item{\code{CA13DD}}{Place where anti-malarials were obtained}
  \item{\code{CA13EE}}{When was anti-malarial first taken}
  \item{\code{CA15}}{What was done to dispose of the stools}
  \item{\code{AN1}}{Measurer's identification code}
  \item{\code{AN2}}{Result of height / length and weight measurement}
  \item{\code{AN3}}{Child's weight (kilograms)}
  \item{\code{AN3A}}{Child undressed to the minimum}
  \item{\code{AN4}}{Child's length or height (centimetres)}
  \item{\code{AN4A}}{Child measured lying or standing}
  \item{\code{HAP}}{Height for age percentile NCHS}
  \item{\code{HAZ}}{Height for age z-score NCHS}
  \item{\code{HAM}}{Height for age percent of reference median NCHS}
  \item{\code{WAP}}{Weight for age percentile NCHS}
  \item{\code{WAZ}}{Weight for age z-score NCHS}
  \item{\code{WAM}}{Weight for age percent of reference median NCHS}
  \item{\code{WHP}}{Weight for height percentile NCHS}
  \item{\code{WHZ}}{Weight for height z-score NCHS}
  \item{\code{WHM}}{Weight for height percent of reference median NCHS}
  \item{\code{FLAG}}{Flag for anthropometric indicators}
  \item{\code{BMI}}{Body Mass Index WHO}
  \item{\code{ZBMI}}{Body Mass Index z-score WHO}
  \item{\code{HAZ2}}{Height for age z-score WHO}
  \item{\code{WAZ2}}{Weight for age z-score WHO}
  \item{\code{WHZ2}}{Weight for height z-score WHO}
  \item{\code{HAZFLAG}}{Height for age flag WHO}
  \item{\code{WAZFLAG}}{Weight for age flag WHO}
  \item{\code{WHZFLAG}}{Weight for height flag WHO}
  \item{\code{BMIFLAG}}{BMI flag WHO}
  \item{\code{WHZNOAGE}}{Weight for height - Age flag WHO}
  \item{\code{HH6}}{Area (Original)}
  \item{\code{HH7}}{District (Original)}
  \item{\code{HL4}}{Sex}
  \item{\code{ED4A}}{Highest level of education attended}
  \item{\code{ED4B}}{Highest grade completed at that level}
  \item{\code{CDOI}}{Date of interview child (CMC)}
  \item{\code{CDOB}}{Date of birth of child (CMC)}
  \item{\code{CAGE}}{Age (months)}
  \item{\code{CAGE.6}}{Age}
  \item{\code{CAGE.11}}{Age}
  \item{\code{CAGED}}{Age in days}
  \item{\code{melevel}}{Mother's education}
  \item{\code{division}}{Division}
  \item{\code{hh7r}}{District}
  \item{\code{hh6r}}{Area}
  \item{\code{suburban}}{}
  \item{\code{stratum}}{Straum (Original)}
  \item{\code{wscore}}{Combined wealth score}
  \item{\code{windex5}}{Wealth index quintile}
  \item{\code{wscoreu}}{Urban wealth score}
  \item{\code{windex5u}}{Urban wealth index quintile}
  \item{\code{wscorer}}{Rural wealth score}
  \item{\code{windex5r}}{Rural wealth index quintile}
  \item{\code{chweight}}{Children's sample weight}
  }}
\usage{
data(MICS2014Ch)
}
\description{
\code{MICS2014Ch} data of Child questionnaire, MICS Punjab, 2014.
}
\examples{
data(MICS2014Ch)
}
\references{
\enumerate{
\item UNICEF MICS (\url{http://www.mics.unicef.org/surveys}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  \item{Muhammad Usman } {(\email{usmann75@hotmail.com})}
  }
}
\keyword{datasets}
